backend/requirements.txt
aiofiles==23.2.1
alembic==1.13.1
asyncio==3.4.3
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.1
bcrypt==4.1.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Cython==3.0.10
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
google-auth==2.30.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
inject==5.2.1
itsdangerous==2.1.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
passlib==1.7.4
provide==0.8
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.26.0
pyppeteer==2.0.0
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.6
rich==13.7.1
rsa==4.9
shellingham==1.5.4
singleton==0.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
uvicorn==0.29.0
watchfiles==0.21.0
fastapi-sso==0.15.0